script, binary: Target 3.65.0 upstream.
* Update upstream from 3.62.2 to 3.65.0
* Explicitly ensure Windows 7 is supported.
* Disable updates.

+++ filezilla.sh
@@ -1,6 +1,13 @@
 #!/bin/sh
 
-# Compile FileZilla for Windows on Ubuntu
+# Compile FileZilla for Windows on Ubuntu 22 or Debian 12 (11 not supported)
+
+START_OVER=1 # Completely start a fresh build from scratch
+INSTALL_PACKAGES=1 # Automatically install required packages. Only needed the first time.
+NEED_BOOST_REGEX=0 # needed for 3.65.0 and newer but not for 3.62.2
+
+# Last tested with FileZilla 3.65.0, Debian 12
+# Originally tested with FileZilla 3.62.2, Ubuntu 22
 
 # Based on (with modifications from):
 # https://svn.filezilla-project.org/filezilla/FileZilla3/
@@ -8,33 +15,64 @@
 
 set -e
 
-rm -rf ~/prefix
-rm -rf ~/src
+if [ "$START_OVER" = "1" ]; then
+	rm -rf ~/prefix
+	rm -rf ~/src
+fi
 
 # These must be run as root
 dpkg --add-architecture i386
-apt update
-apt install -y automake autoconf libtool make gettext lzip
-apt install -y mingw-w64 pkg-config wx-common wine wine64 wine32 wine-binfmt subversion git
+
+if [ "$INSTALL_PACKAGES" = "1" ]; then
+	apt update
+	apt install -y automake autoconf libtool make gettext lzip
+	apt install -y mingw-w64 pkg-config wx-common wine wine64 wine32 wine-binfmt subversion git g++
+fi
 
 # Don't install packages, e.g. libfilezilla-dev for cross-compiling!
 
 # Everything else can be run as non-root, if desired.
-mkdir ~/prefix
-mkdir ~/src
+
+if [ "$START_OVER" = "1" ]; then
+	mkdir ~/prefix
+	mkdir ~/src
+fi
+
 export PATH="$HOME/prefix/bin:$PATH"
 export LD_LIBRARY_PATH="$HOME/prefix/lib:$LD_LIBRARY_PATH"
 export PKG_CONFIG_PATH="$HOME/prefix/lib/pkgconfig:$PKG_CONFIG_PATH"
 export TARGET_HOST=x86_64-w64-mingw32
 
+# Explicitly target Windows 7 to avoid using GetSystemTimePreciseAsFileTime, which is Windows 8+ only.
+# See also: --with-default-win32-winnt=0x0601
+export CPPFLAGS="-D_WIN32_WINNT=0x0601"
+
 wine reg add HKCU\\Environment /f /v PATH /d "`x86_64-w64-mingw32-g++ -print-search-dirs | grep ^libraries | sed 's/^libraries: =//' | sed 's/:/;z:/g' | sed 's/^\\//z:\\\\\\\\/' | sed 's/\\//\\\\/g'`"
+
+if [ "$NEED_BOOST_REGEX" = "1" ]; then
+	# Boost Regex 1.76+
+	cd ~/src
+	wget https://boostorg.jfrog.io/artifactory/main/release/1.76.0/source/boost_1_76_0.tar.bz2
+	tar --bzip2 -xf boost_1_76_0.tar.bz2
+	cd boost_1_76_0
+	# The following two steps *might* be optional? Didn't work the first time, but worked eventually...
+	#./bootstrap.sh --prefix="$HOME/prefix"
+	#./b2 headers
+
+	# Determine the include path from the first line of:
+	x86_64-w64-mingw32-g++ -std=c++17 -c -std=c++17 -E -x c++ -v /dev/null
+
+	cp -r ~/src/boost_1_76_0/boost/ /usr/lib/gcc/x86_64-w64-mingw32/12-win32/include/c++/
+fi
 
 # GMP
 cd ~/src
-wget https://gmplib.org/download/gmp/gmp-6.2.1.tar.lz
-tar xf gmp-6.2.1.tar.lz
+if [ ! -d gmp-6.2.1 ]; then
+	wget https://gmplib.org/download/gmp/gmp-6.2.1.tar.lz
+	tar xf gmp-6.2.1.tar.lz
+fi
 cd gmp-6.2.1
-CC_FOR_BUILD=gcc ./configure --host=$TARGET_HOST --prefix="$HOME/prefix" --disable-static --enable-shared --enable-fat
+CC_FOR_BUILD=gcc ./configure --with-default-win32-winnt=0x0601 --host=$TARGET_HOST --prefix="$HOME/prefix" --disable-static --enable-shared --enable-fat
 # This fails to build on Debian 11 for me, but it works on Ubuntu
 make
 make install
@@ -44,7 +82,7 @@ cd ~/src
 wget https://ftp.gnu.org/gnu/nettle/nettle-3.7.3.tar.gz
 tar xf nettle-3.7.3.tar.gz
 cd nettle-3.7.3
-./configure --host=$TARGET_HOST --prefix="$HOME/prefix" --enable-shared --disable-static --enable-fat LDFLAGS="-L$HOME/prefix/lib" CPPFLAGS="-I$HOME/prefix/include"
+./configure --with-default-win32-winnt=0x0601 --host=$TARGET_HOST --prefix="$HOME/prefix" --enable-shared --disable-static --enable-fat LDFLAGS="-L$HOME/prefix/lib" CPPFLAGS="-I$HOME/prefix/include -D_WIN32_WINNT=0x0601"
 make
 make install
 
@@ -53,7 +91,7 @@ cd ~/src
 wget https://www.gnupg.org/ftp/gcrypt/gnutls/v3.7/gnutls-3.7.8.tar.xz
 tar xvf gnutls-3.7.8.tar.xz
 cd gnutls-3.7.8
-./configure --host=$TARGET_HOST --prefix="$HOME/prefix" --enable-shared --disable-static --without-p11-kit --with-included-libtasn1 --with-included-unistring --enable-local-libopts --disable-srp-authentication --disable-dtls-srtp-support --disable-heartbeat-support --disable-psk-authentication --disable-anon-authentication --disable-openssl-compatibility --without-tpm --without-zlib --without-brotli --without-zstd --without-idn --enable-threads=windows --disable-cxx LDFLAGS="-L$HOME/prefix/lib" CPPFLAGS="-I$HOME/prefix/include"
+./configure --with-default-win32-winnt=0x0601 --host=$TARGET_HOST --prefix="$HOME/prefix" --enable-shared --disable-static --without-p11-kit --with-included-libtasn1 --with-included-unistring --enable-local-libopts --disable-srp-authentication --disable-dtls-srtp-support --disable-heartbeat-support --disable-psk-authentication --disable-anon-authentication --disable-openssl-compatibility --without-tpm --without-zlib --without-brotli --without-zstd --without-idn --enable-threads=windows --disable-cxx LDFLAGS="-L$HOME/prefix/lib" CPPFLAGS="-I$HOME/prefix/include -D_WIN32_WINNT=0x0601"
 
 # Could fail at some point, but if we get past lib, that's all we need, we're okay
 (make && make install) || (make || (cd lib && make install))
@@ -62,10 +100,11 @@ cd gnutls-3.7.8
 cd ~/src
 #wget https://sqlite.org/2018/sqlite-autoconf-32600-00.tar.gz # 
 #wget https://www.sqlite.com/matrix/2022/sqlite-autoconf-3390400.tar.gz # correct link, but server is down
-wget https://github.com/elimuhubconsultant-co-ke/armbian_build/blob/master/sqlite-autoconf-3390300.tar.gz?raw=true
-tar xvzf sqlite-autoconf-3390300.tar.gz?raw=true
+wget https://github.com/elimuhubconsultant-co-ke/armbian_build/raw/master/sqlite-autoconf-3390300.tar.gz
+tar xvzf sqlite-autoconf-3390300.tar.gz
 cd sqlite-autoconf-3390300
-./configure --host=$TARGET_HOST --prefix="$HOME/prefix" --enable-shared --disable-static --disable-dynamic-extensions
+make clean
+./configure --with-default-win32-winnt=0x0601 --host=$TARGET_HOST --prefix="$HOME/prefix" --enable-shared --disable-static --disable-dynamic-extensions
 make
 make install
 
@@ -80,20 +119,21 @@ wine nsis-3.04-setup.exe /S
 cd ~/src
 git clone --branch WX_3_0_BRANCH --single-branch https://github.com/wxWidgets/wxWidgets.git wx3
 cd wx3
-wget https://filezilla-project.org/nightlies/2023-01-17/patches/wx3/cross_compiling.patch
-git apply cross_compiling.patch
-./configure --host=$TARGET_HOST --prefix="$HOME/prefix" --enable-shared --disable-static --enable-gui --enable-printfposparam
+#wget https://filezilla-project.org/nightlies/2023-01-17/patches/wx3/cross_compiling.patch
+#git apply cross_compiling.patch
+./configure --with-default-win32-winnt=0x0601 --host=$TARGET_HOST --prefix="$HOME/prefix" --enable-shared --disable-static --enable-gui --enable-printfposparam
 make
 make install
 cp $HOME/prefix/lib/wx*.dll $HOME/prefix/bin
 
 # libfilezilla
+export CXXFLAGS=-std=c++17
 cd ~/src
 #svn co https://svn.filezilla-project.org/svn/libfilezilla/tags/0.33.0 lfz
 svn co https://svn.filezilla-project.org/svn/libfilezilla/trunk lfz
 cd lfz
 autoreconf -i
-./configure --host=$TARGET_HOST --prefix="$HOME/prefix" --enable-shared --disable-static
+./configure --with-default-win32-winnt=0x0601 --host=$TARGET_HOST --prefix="$HOME/prefix" --enable-shared --disable-static
 make
 make install
 
@@ -107,10 +147,20 @@ cd fz
 wget https://raw.githubusercontent.com/InterLinked1/fastfilezilla/master/anytimeouts.diff
 svn patch anytimeouts.diff
 
+#autoupdate
 autoreconf -i
 # https://forum.filezilla-project.org/viewtopic.php?style=246&t=42578
 export PKG_CONFIG_PATH="$HOME/prefix/lib/pkgconfig/"
-./configure --host=$TARGET_HOST --prefix="$HOME/prefix" --enable-shared --disable-static --with-pugixml=builtin
+
+# If configure fails, for debugging, these SHOULD succeed. If not, boost headers aren't installed.
+# echo "#include <boost/regex.hpp>" > conftest.cpp
+# x86_64-w64-mingw32-g++ -std=c++17 -c -std=c++17 -Wall -g conftest.cpp
+
+# Do not run make clean here. There is no clean target.
+
+# Disable updates, since we're compiling a custom build, so those aren't relevant. We'll need to rebuild from source again, on our terms.
+# --disable-manualupdatecheck includes --disable-autoupdatecheck
+./configure --with-default-win32-winnt=0x0601 --host=$TARGET_HOST --prefix="$HOME/prefix" --enable-shared --disable-static --with-pugixml=builtin --disable-manualupdatecheck
 # NOTE: If you make changes in src, run make clean first, not just make!
 make
 # strip debug symbols
@@ -119,7 +169,13 @@ $TARGET_HOST-strip src/putty/.libs/fzsftp.exe
 $TARGET_HOST-strip src/putty/.libs/fzputtygen.exe
 $TARGET_HOST-strip src/fzshellext/64/.libs/libfzshellext-0.dll
 $TARGET_HOST-strip src/fzshellext/32/.libs/libfzshellext-0.dll
-$TARGET_HOST-strip data/dlls/*.dll
+#$TARGET_HOST-strip data/dlls/*.dll
+# Seems the folder name got changed?
+$TARGET_HOST-strip data/dlls/*.dll || $TARGET_HOST-strip data/dlls_gui/*.dll
+
+# This symbol is incompatible with Windows 7 and must not exist in the result
+grep -R "GetSystemTimePreciseAsFileTime"
+
 cd data
 
 # This is slightly modified from the original command on the website
@@ -127,5 +183,6 @@ cd data
 wine "$HOME/.wine/drive_c/Program Files (x86)/NSIS/makensis.exe" install.nsi
 
 # Voila, there's now FileZilla_3_setup.exe in the current directory.
-printf "Cross-compilation of FileZilla has finished."
+printf "Cross-compilation of FileZilla has finished.\n"
+pwd
 ls -la FileZilla_3_setup.exe
